mbsync: PassCmd, separate encrypted file
`pass` should not be used to recover passwords with mbsync. In fact, the
program will most likely use `gpg-agent` and maximize (default: 2 hours)
the caching time of passwords, as per gpg-agent(1). A different key with
different password should be used in order to assure best security
conditions of the password store.