qutebrowser: remap (inc/dec)rement of URL no, g-+
The default ctrl-x, ctrl-a are too dangerous during rapid manipulations
of text since ctrl-x can lead to leaving the page while you're filling
important form for instance. This occur when one thinks of using ctrl-x
for cutting text in a text box, but doing this manipulation too quickly
in normal mode ends up in a big mess.