vim/fugitive: [/] get rid of system() bad EOL char
The function system() seems to include a character '\r' which was
directly put into the command to call, hence creating an inappropriate
behaviour. This is fixed with the usage of systemlist(...)[0]. The usage
of `cut` has been replaced with VimL's substitute in order to keep most
of the text part of the running language.