vim: [+] fzf plugin, [-] ctrl-p
ctrl-p is a simple Vim plugin is quite limited compared to fzf. The
latter is an easily pluggable utility for which there exist a Vim plugin
with extensive number of features that are all directly plugged from the
command line interface. In short: fzf has a better design more features.

See the links below:

    https://github.com/junegunn/fzf
    https://github.com/junegunn/fzf.vim