task: [-] toto report, scheduled:someday
Don't consider tasks scheduled someday as being "todo". Instead consider
tasks scheduled before 1 week later from today. When I want to schedule
a task some time soon, in reality what I mean is that I want to schedule
it now because it could be done any time soon. It is on the priority
"todo list". If I want to remove it from the list, then I schedule it
someday (with `scheduled:someday`) and then it won't appear in the todo
report.